+++ example/lithia.config.js
@@ -1,5 +1,0 @@
-import { defineLithiaConfig } from '../dist';
-
-export default defineLithiaConfig({
-  globalMiddlewares: [],
-});