+++ examples/1-basic-api/lithia.config.js
@@ -1,0 +1,5 @@
+const { defineLithiaConfig } = require('lithia');
+
+module.exports = defineLithiaConfig({
+  globalMiddlewares: [],
+});