+++ examples/1-basic-api/src/routes/hello.get.ts
@@ -1,0 +1,7 @@
+import { LithiaRequest, LithiaResponse } from 'lithia';
+
+export default async function handler(req: LithiaRequest, res: LithiaResponse) {
+  res.json({
+    message: 'Hello, from Lithia! 🚀🚀🚀🚀',
+  });
+}